+++ test/jarbler/config_test.rb
@@ -1,8 +1,10 @@
 require 'minitest/autorun'
 require 'jarbler/config'
+require 'test_helper'
 
 class ConfigTest < Minitest::Test
   def setup
+    super
     @config = Jarbler::Config.new
   end